#12 Rename folder 'modal' to 'modals
